requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
PyYAML==6.0.1
watchdog==3.0.0
pytz==2023.3
requests==2.31.0
